Prevent checkbox facets from breaking the search/filtering
When you have a checkbox value in the list of facets, the filtering no longer works because this class converts true and false to a numerical value. But the numerical value does not match elasticsearch's expectation. With this addition to the mapper, this will now work as expected; though the label still shows an ugly label ;)

This will fix https://github.com/BitBagCommerce/SyliusElasticsearchPlugin/issues/113